+++ app/api/Search.scala
@@ -1,9 +1,11 @@
 package api
 
-import services.{RdfSPARQLService, DatasetService, FileService, CollectionService, PreviewService, MultimediaQueryService, ElasticsearchPlugin}
+import api.Permission._
+import services.{RdfSPARQLService, DatasetService, FileService, CollectionService, PreviewService, SpaceService,
+MultimediaQueryService, ElasticsearchPlugin}
 import play.Logger
 import scala.collection.mutable.{ListBuffer, HashMap}
-import util.SearchResult
+import util.{SearchUtils, SearchResult}
 import play.api.libs.json.{JsObject, Json, JsValue}
 import play.api.libs.json.Json.toJson
 import javax.inject.{Inject, Singleton}
@@ -18,55 +20,42 @@ class Search @Inject() (
    collections: CollectionService,
    previews: PreviewService,
    queries: MultimediaQueryService,
+   spaces: SpaceService,
    sparql: RdfSPARQLService)  extends ApiController {
 
   /** Search using a simple text string with filters */
-  def search(query: String, resource_type: Option[String],
-             datasetid: Option[String], collectionid: Option[String], spaceid: Option[String], folderid: Option[String],
-             field: Option[String], tag: Option[String]) = PermissionAction(Permission.ViewDataset) { implicit request =>
+  def search(query: String, resource_type: Option[String], datasetid: Option[String], collectionid: Option[String],
+             spaceid: Option[String], folderid: Option[String], field: Option[String], tag: Option[String],
+             from: Option[Int], size: Option[Int], page: Option[Int]) = PermissionAction(Permission.ViewDataset) { implicit request =>
     current.plugin[ElasticsearchPlugin] match {
       case Some(plugin) => {
-        val response = plugin.searchWithParameters(query, resource_type, datasetid, collectionid, spaceid, folderid, field, tag)
-
-        var filesFound = ListBuffer.empty[UUID]
-        var datasetsFound = ListBuffer.empty[UUID]
-        var collectionsFound = ListBuffer.empty[UUID]
-
-        for (resource <- response) {
-          resource.resourceType match {
-            case ResourceRef.file => filesFound += resource.id
-            case ResourceRef.dataset => datasetsFound += resource.id
-            case ResourceRef.collection => collectionsFound += resource.id
-            case other => Logger.debug(s"search result resource type not supported: ${other.toString}")
+        // If from is specified, use it. Otherwise use page * size of page if possible, otherwise use 0.
+        val from_index = from match {
+          case Some(f) => from
+          case None => page match {
+            case Some(p) => Some(size.getOrElse(0) * p)
+            case None => None
           }
         }
 
-        val filesList = files.get(filesFound.toList).map(f => toJson(f))
-        val datasetsList = datasets.get(datasetsFound.toList).map(ds => toJson(ds))
-        val collectionsList = collections.get(collectionsFound.toList).map(c => toJson(c))
+        // TODO: Better way to build a URL?
+        val source_url = s"/api/search?query=$query" +
+          (resource_type match {case Some(x) => s"&resource_type=$x" case None => ""}) +
+          (datasetid match {case Some(x) => s"&datasetid=$x" case None => ""}) +
+          (collectionid match {case Some(x) => s"&collectionid=$x" case None => ""}) +
+          (spaceid match {case Some(x) => s"&spaceid=$x" case None => ""}) +
+          (folderid match {case Some(x) => s"&folderid=$x" case None => ""}) +
+          (field match {case Some(x) => s"&field=$x" case None => ""}) +
+          (tag match {case Some(x) => s"&tag=$x" case None => ""})
 
-        resource_type match {
-          case Some("file") => Ok(toJson(Map[String, JsValue]("files" -> toJson(filesList))))
-          case Some("dataset") => Ok(toJson(Map[String, JsValue]("datasets" -> toJson(datasetsList))))
-          case Some("collection") => Ok(toJson(Map[String, JsValue]("collections" -> toJson(collectionsList))))
+        // Add space filter to search here as a simple permissions check
+        val permitted = spaces.listAccess(0, Set[Permission](Permission.ViewSpace), request.user, true, true, false, false).map(sp => sp.id)
 
-          case _ => {
-            // If datasetid is provided, only files are returned
-            datasetid match {
-              case Some(dsid) => Ok(toJson(Map[String, JsValue](
-                "files" -> toJson(filesList)
-              )))
-              case None => {
-                // collection and space IDs do not restrict resource type
-                Ok(toJson(Map[String, JsValue](
-                  "files" -> toJson(filesList),
-                  "datasets" -> toJson(datasetsList),
-                  "collections" -> toJson(collectionsList)
-                )))
-              }
-            }
-          }
-        }
+        val response = plugin.search(query, resource_type, datasetid, collectionid, spaceid, folderid, field, tag, from_index, size, permitted, request.user)
+
+        val result = SearchUtils.prepareSearchResponse(response, source_url, request.user)
+        // TODO: Post-process this on batch query
+        Ok(toJson(result))
       }
       case None => {
         Logger.debug("Search plugin not enabled")
@@ -83,32 +72,13 @@ class Search @Inject() (
       current.plugin[ElasticsearchPlugin] match {
         case Some(plugin) => {
           val queryList = Json.parse(query).as[List[JsValue]]
-          val results = plugin.search(queryList, grouping, from, size)
+          val response = plugin.search(queryList, grouping, from, size, user)
 
-          var filesFound = ListBuffer.empty[UUID]
-          var datasetsFound = ListBuffer.empty[UUID]
-          var collectionsFound = ListBuffer.empty[UUID]
+          // TODO: Better way to build a URL?
+          val source_url = s"/api/search?query=$query&grouping=$grouping"
 
-          for (resource <- results) {
-            resource.resourceType match {
-              case ResourceRef.file => filesFound += resource.id
-              case ResourceRef.dataset => datasetsFound += resource.id
-              case ResourceRef.collection => collectionsFound += resource.id
-              case _ => {}
-            }
-          }
-
-          val filesList = files.get(filesFound.toList).map(f => toJson(f))
-          val datasetsList = datasets.get(datasetsFound.toList).map(ds => toJson(ds))
-          val collectionsList = collections.get(collectionsFound.toList).map(c => toJson(c))
-
-          // Use "distinct" to remove duplicate results.
-          Ok(JsObject(Seq(
-            "datasets" -> toJson(datasetsList.distinct),
-            "files" -> toJson(filesList.distinct),
-            "collections" -> toJson(collectionsList.distinct),
-            "count" -> toJson(results.length)
-          )))
+          val result = SearchUtils.prepareSearchResponse(response, source_url, user)
+          Ok(toJson(result))
         }
         case None => {
           BadRequest("Elasticsearch plugin could not be reached")
@@ -135,8 +105,7 @@ class Search @Inject() (
   /**
    * Search MultimediaFeatures.
    */
-  def searchMultimediaIndex(section_id: UUID) = PermissionAction(Permission.ViewSection,
-    Some(ResourceRef(ResourceRef.section, section_id))) {
+  def searchMultimediaIndex(section_id: UUID) = PermissionAction(Permission.ViewSection, Some(ResourceRef(ResourceRef.section, section_id))) {
     implicit request =>
 
     // Finding IDs of spaces that the user has access to