+++ app/models/Dataset.scala
@@ -61,7 +61,8 @@ object Dataset {
         "created" -> dataset.created.toString,
         "thumbnail" -> datasetThumbnail,
         "authorId" -> dataset.author.id,
-        "spaces" -> dataset.spaces)
+        "spaces" -> dataset.spaces,
+        "resource_type" -> "dataset")
     }
   }
 }